re-organize the code and structure
