+++ src/annotate/__init__.py
@@ -1,0 +1,3 @@
+from .annotation_pipeline import LLMAnnotationPipeline, AnnotationPromptGenerator
+
+__all__ = ['LLMAnnotationPipeline', 'AnnotationPromptGenerator']
