+++ src/evaluate/__init__.py
@@ -1,0 +1,4 @@
+from .evaluator import ScamDetectionEvaluator, PhishingEvaluator
+from .prompt_generator import PromptGenerator
+
+__all__ = ['ScamDetectionEvaluator', 'PhishingEvaluator', 'PromptGenerator']
