+++ src/evaluate/evaluator.py
@@ -1,108 +1,22 @@
-#!/usr/bin/env python3
-"""
-LLM Annotation Pipeline for Scam Detection
-
-This pipeline uses an LLM to annotate data by providing explanations for why content
-is labeled as legitimate or scam. The LLM is given both the original content and 
-the correct answer, then asked to explain the reasoning behind the classification.
-
-Usage:
-    python annotation_pipeline.py --dataset path/to/dataset.csv --provider openai --model gpt-4 --sample-size 100
-"""
-
-import argparse
-import sys
-import os
-from pathlib import Path
-from typing import List, Dict, Any, Optional
 import pandas as pd
+from typing import List, Dict, Any, Optional
+from src.llm_core.api_provider import LLM
+from src.llm_core.api_call import make_api_call, parse_structured_output, make_api_call_async, parse_structured_output_async
+from src.utility.data_loader import DatasetLoader
+from src.evaluate.prompt_generator import PromptGenerator
+from src.utility.metrics_calculator import MetricsCalculator
+from src.utility.results_saver import ResultsSaver
 import asyncio
 import time
 import json
 from datetime import datetime
+from pathlib import Path
 from tqdm import tqdm
 
-# Add current directory to Python path to allow imports
-current_dir = Path(__file__).parent
-sys.path.insert(0, str(current_dir))
-
-from api_provider import LLM
-from api_call import make_api_call, parse_structured_output, make_api_call_async, parse_structured_output_async
-from data_loader import DatasetLoader
-from results_saver import ResultsSaver
-from pydantic import BaseModel
-
-class AnnotationResponseSchema(BaseModel):
-    explanation: str
-    key_indicators: List[str]
-    confidence: str  # "high", "medium", "low"
-    usability: bool  # True if data is meaningful, False if not meaningful/nonsensical
-
-class AnnotationPromptGenerator:
-    """Generate prompts for annotation tasks"""
-    
-    def __init__(self, features: List[str], content_columns: List[str] = None):
-        self.features = features
-        self.content_columns = content_columns or features
-    
-    def get_system_prompt(self) -> str:
-        """Generate system prompt for annotation"""
-        return """
-You are an expert cybersecurity analyst and educator specializing in scam detection.
-Your task is to provide detailed educational explanations for why content is classified 
-as either a scam or legitimate.
-
-You will be given:
-1. The original content (email, message, etc.)
-2. The correct classification (scam or legitimate)
-
-Your job is to explain WHY this classification is correct by:
-- Identifying key indicators that support the classification
-- Explaining the reasoning in an educational manner
-- Highlighting specific techniques or patterns used (for scams) or trust signals (for legitimate content)
-- Providing insights that would help someone learn to identify similar cases
-- Assessing whether the content is meaningful and usable for analysis (**Do not include explanation for usability in your finalresponse**)
-
-Guidelines:
-- Be thorough but concise in your explanations
-- Focus on actionable insights that help with detection
-- For scams: explain the deception techniques, red flags, and malicious intent
-- For legitimate content: explain trust indicators and why it's genuine
-- Always be educational and help build detection skills
-- Evaluate if the content makes sense and is meaningful for analysis
-
-Respond with:
-- explanation: A detailed explanation of why this content has the given classification
-- key_indicators: A list of specific indicators/features that support the classification
-- confidence: Your confidence level in the explanation ("high", "medium", "low")
-- usability: True if the content is meaningful and makes sense for analysis, False if the content is nonsensical, corrupted, or not meaningful
-"""
-    
-    def create_annotation_prompt(self, row: Dict[str, Any], correct_label: str) -> str:
-        """Create annotation prompt with content and correct answer"""
-        classification = "SCAM" if correct_label == "1" or correct_label == 1 else "LEGITIMATE"
-        
-        prompt_parts = [
-            f"Please analyze the following content that is classified as {classification}.\n",
-            "Provide a detailed explanation of why this classification is correct.\n\n",
-            "CONTENT TO ANALYZE:\n"
-        ]
-        
-        # Add content features
-        for feature in self.content_columns:
-            if feature in row:
-                value = row.get(feature, "")
-                if value and str(value).strip():
-                    prompt_parts.append(f"{feature}: {str(value).strip()}\n")
-        
-        prompt_parts.append(f"\nCORRECT CLASSIFICATION: {classification}\n")
-        prompt_parts.append("\nProvide your detailed explanation of why this classification is correct.")
-        
-        return "".join(prompt_parts)
-
-class LLMAnnotationPipeline:
+class ScamDetectionEvaluator:
     """
-    Main annotation pipeline that generates explanations for labeled data
+    Main evaluator class that orchestrates the entire scam detection evaluation pipeline.
+    Designed to work with any dataset that has a 'label' column and text content.
     """
     
     def __init__(self, 
@@ -112,24 +26,21 @@ class LLMAnnotationPipeline:
                  sample_size: int = 100,
                  balanced_sample: bool = False,
                  random_state: int = 42,
-                 content_columns: Optional[List[str]] = None,
-                 output_dir: str = "annotations",
                  enable_thinking: bool = False,
+                 content_columns: Optional[List[str]] = None,
                  use_structure_model: bool = False):
         """
-        Initialize the annotation pipeline
+        Initialize the evaluator with dataset and model configuration
         
         Args:
             dataset_path: Path to the CSV dataset file
             provider: LLM provider (e.g., 'openai', 'anthropic', 'local')
             model: Model name
-            sample_size: Number of samples to annotate
+            sample_size: Number of samples to evaluate
             balanced_sample: Whether to sample equal numbers of scam and legitimate messages
             random_state: Random seed for reproducibility
-            content_columns: List of column names to use as content for annotation
-            output_dir: Directory to save annotation results
-            enable_thinking: Whether to enable thinking tokens in prompts
-            use_structure_model: Whether to use a separate structure model for parsing
+            content_columns: List of column names to use as content for evaluation.
+                           If None, uses all non-label columns.
         """
         self.dataset_path = dataset_path
         self.provider = provider
@@ -138,9 +49,6 @@ class LLMAnnotationPipeline:
         self.balanced_sample = balanced_sample
         self.random_state = random_state
         self.content_columns = content_columns
-        self.output_dir = output_dir
-        self.enable_thinking = enable_thinking
-        self.use_structure_model = use_structure_model
         
         # Initialize components
         self.data_loader = DatasetLoader(dataset_path)
@@ -148,7 +56,9 @@ class LLMAnnotationPipeline:
         self.llm = None
         self.structure_model = None
         self.prompt_generator = None
-        self.annotations = []
+        self.results = []
+        self.use_structure_model = use_structure_model
+        self.enable_thinking = enable_thinking
         
         # Checkpoint state
         self.current_index = 0
@@ -169,7 +79,7 @@ class LLMAnnotationPipeline:
             raise Exception(f"Error initializing LLM: {e}")
     
     def load_and_prepare_data(self) -> pd.DataFrame:
-        """Load dataset and prepare sample for annotation"""
+        """Load dataset and prepare sample for evaluation"""
         # Load full dataset
         self.data_loader.load_dataset()
         
@@ -189,24 +99,29 @@ class LLMAnnotationPipeline:
         else:
             sample_df = self.data_loader.sample_data(self.sample_size, self.random_state)
         
-        # Initialize prompt generator
-        self.prompt_generator = AnnotationPromptGenerator(self.data_loader.features, self.content_columns)
+        # Initialize prompt generator with available features and specified content columns
+        self.prompt_generator = PromptGenerator(self.data_loader.features, self.content_columns)
+        
+        # Calculate sample distribution
+        sample_scam = len(sample_df[sample_df['label'] == 1])
+        sample_legit = len(sample_df[sample_df['label'] == 0])
         
         print(f"\nDataset: {self.data_loader.dataset_name}")
         print(f"Sample size: {len(sample_df)} records")
-        print(f"Content features: {', '.join(self.content_columns)}")
+        print(f"Sample distribution: {sample_scam} scam, {sample_legit} legitimate")
+        print(self.prompt_generator.get_features_summary())
         
         return sample_df
     
-    def annotate_sample(self, sample_df: pd.DataFrame) -> List[Dict[str, Any]]:
+    def evaluate_sample(self, sample_df: pd.DataFrame) -> List[Dict[str, Any]]:
         """
-        Annotate the sample dataset using the LLM
+        Evaluate the sample dataset using the LLM
         
         Args:
-            sample_df: Sample dataframe to annotate
+            sample_df: Sample dataframe to evaluate
             
         Returns:
-            List of annotation results
+            List of evaluation results
         """
         if self.llm is None:
             raise ValueError("LLM not initialized. Call setup_llm() first.")
@@ -214,64 +129,74 @@ class LLMAnnotationPipeline:
         if self.prompt_generator is None:
             raise ValueError("Prompt generator not initialized. Call load_and_prepare_data() first.")
         
-        annotations = []
+        results = []
         system_prompt = self.prompt_generator.get_system_prompt()
         
         print("\n" + "="*80)
-        print("STARTING LLM ANNOTATION PROCESS")
+        print("STARTING SCAM DETECTION EVALUATION")
         print("="*80)
         
         # Use tqdm for progress tracking
-        for i, (_, row) in enumerate(tqdm(sample_df.iterrows(), total=len(sample_df), desc="Annotating")):
+        for i, (_, row) in enumerate(tqdm(sample_df.iterrows(), total=len(sample_df), desc="Evaluating")):
             
-            # Create annotation prompt with correct label
-            user_prompt = self.prompt_generator.create_annotation_prompt(row.to_dict(), str(row['label']))
+            # Create user prompt with specified content features
+            user_prompt = self.prompt_generator.create_user_prompt(row.to_dict())
             
             try:
                 if self.use_structure_model:
                     # Make API call
                     response = make_api_call(self.llm, system_prompt, user_prompt, enable_thinking=self.enable_thinking, structure_model=True)
+                    
                     # Remove thinking tokens if they exist
                     if self.enable_thinking or ('<think>' in response and '</think>' in response):
                         import re
                         # Remove everything between <think> and </think> tags
                         response = re.sub(r'<think>.*?</think>', '', response, flags=re.DOTALL).strip()
-                    response = parse_structured_output(self.structure_model, response, AnnotationResponseSchema)
+                        # Response cleaned of thinking tokens
+                    
+                    response = parse_structured_output(self.structure_model, response)
+                    # Response parsed successfully
                 else:
-                    # Make API call with custom structured output for annotations
-                    prompt_template = self._get_annotation_prompt_template()
-                    messages = prompt_template.invoke({"system_prompt": system_prompt, "user_prompt": user_prompt})
-                    client = self.llm.with_structured_output(AnnotationResponseSchema)
-                    response = client.invoke(messages)
+                    # Make API call
+                    response = make_api_call(self.llm, system_prompt, user_prompt, structure_model=False)
                 
-                # Create comprehensive annotation record
-                annotation = self._create_annotation_record(i+1, row, response)
-                annotations.append(annotation)
+                # Extract prediction
+                predicted_scam = response.Phishing  # Note: API still uses "Phishing" key for compatibility
+                predicted_label = 1 if predicted_scam else 0
+                actual_label = int(row['label'])  # Should be safe since data is pre-cleaned
                 
-                actual_class = 'Scam' if row['label'] == 1 else 'Legitimate'
-                usability_status = 'Usable' if response.usability else 'Not usable'
-                tqdm.write(f"  Record {i+1}: Class={actual_class}, Confidence={response.confidence}, "
-                          f"Usability={usability_status}, Indicators={len(response.key_indicators)}")
+                # Calculate if prediction is correct
+                is_correct = predicted_label == actual_label
+                
+                # Create comprehensive result record
+                result = self._create_result_record(row, predicted_label, is_correct, response.Reason)
+                
+                results.append(result)
+                
+                # Use tqdm.write to avoid interfering with progress bar
+                tqdm.write(f"  Record {i+1}: Actual={'Scam' if actual_label == 1 else 'Legitimate'}, "
+                          f"Predicted={'Scam' if predicted_label == 1 else 'Legitimate'}, "
+                          f"Correct={is_correct}")
                 
             except Exception as e:
-                tqdm.write(f"  Error annotating record {i+1}: {e}")
-                annotation = self._create_error_annotation_record(i+1, row, str(e))
-                annotations.append(annotation)
+                tqdm.write(f"  Error processing record {i+1}: {e}")
+                result = self._create_error_result_record(row, str(e))
+                results.append(result)
         
-        self.annotations = annotations
-        print(f"\nAnnotation completed. Processed {len(annotations)} records.")
-        return annotations
+        self.results = results
+        print(f"\nEvaluation completed. Processed {len(results)} records.")
+        return results
 
-    async def annotate_sample_async(self, sample_df: pd.DataFrame, concurrent_requests: int = 10) -> List[Dict[str, Any]]:
+    async def evaluate_sample_async(self, sample_df: pd.DataFrame, concurrent_requests: int = 10) -> List[Dict[str, Any]]:
         """
-        Annotate the sample dataset using the LLM with concurrent requests
+        Evaluate the sample dataset using the LLM with concurrent requests
         
         Args:
-            sample_df: Sample dataframe to annotate
+            sample_df: Sample dataframe to evaluate
             concurrent_requests: Number of concurrent requests to make (default: 10)
             
         Returns:
-            List of annotation results
+            List of evaluation results
         """
         if self.llm is None:
             raise ValueError("LLM not initialized. Call setup_llm() first.")
@@ -282,52 +207,63 @@ class LLMAnnotationPipeline:
         system_prompt = self.prompt_generator.get_system_prompt()
         
         print("\n" + "="*80)
-        print("STARTING LLM ANNOTATION PROCESS (ASYNC)")
+        print("STARTING SCAM DETECTION EVALUATION (ASYNC)")
         print(f"Concurrent requests: {concurrent_requests}")
         print("="*80)
         
         # Create semaphore to limit concurrent requests
         semaphore = asyncio.Semaphore(concurrent_requests)
         
-        async def annotate_single_record(i: int, row: pd.Series) -> Dict[str, Any]:
+        # Create progress bar for async processing
+        pbar = tqdm(total=len(sample_df), desc="Evaluating (async)")
+        
+        async def evaluate_single_record(i: int, row: pd.Series) -> Dict[str, Any]:
             async with semaphore:
-                print(f"Annotating record {i+1}/{len(sample_df)}...")
+                # Update progress bar description
+                pbar.set_description(f"Evaluating record {i+1}/{len(sample_df)}")
                 
-                # Create annotation prompt with correct label
-                user_prompt = self.prompt_generator.create_annotation_prompt(row.to_dict(), str(row['label']))
+                # Create user prompt with specified content features
+                user_prompt = self.prompt_generator.create_user_prompt(row.to_dict())
                 
                 try:
                     if self.use_structure_model:
                         # Make async API call
                         response = await make_api_call_async(self.llm, system_prompt, user_prompt, 
                                                            enable_thinking=self.enable_thinking, structure_model=True)
+                        
                         # Remove thinking tokens if they exist
                         if self.enable_thinking or ('<think>' in response and '</think>' in response):
                             import re
                             # Remove everything between <think> and </think> tags
                             response = re.sub(r'<think>.*?</think>', '', response, flags=re.DOTALL).strip()
-                        response = await parse_structured_output_async(self.structure_model, response, AnnotationResponseSchema)
+                            # Response cleaned of thinking tokens
+                        
+                        response = await parse_structured_output_async(self.structure_model, response)
+                        # Response parsed successfully
                     else:
-                        # Make async API call with custom structured output for annotations
-                        prompt_template = self._get_annotation_prompt_template()
-                        messages = prompt_template.invoke({"system_prompt": system_prompt, "user_prompt": user_prompt})
-                        client = self.llm.with_structured_output(AnnotationResponseSchema)
-                        response = await client.ainvoke(messages)
+                        # Make async API call
+                        response = await make_api_call_async(self.llm, system_prompt, user_prompt, structure_model=False)
                     
-                    # Create comprehensive annotation record
-                    annotation = self._create_annotation_record(i+1, row, response)
+                    # Extract prediction
+                    predicted_scam = response.Phishing  # Note: API still uses "Phishing" key for compatibility
+                    predicted_label = 1 if predicted_scam else 0
+                    actual_label = int(row['label'])  # Should be safe since data is pre-cleaned
                     
-                    actual_class = 'Scam' if row['label'] == 1 else 'Legitimate'
-                    usability_status = 'Usable' if response.usability else 'Not usable'
-                    tqdm.write(f"Record {i+1} - Class: {actual_class}, Confidence: {response.confidence}, "
-                              f"Usability: {usability_status}, Indicators: {len(response.key_indicators)}")
+                    # Calculate if prediction is correct
+                    is_correct = predicted_label == actual_label
                     
-                    return annotation
+                    # Create comprehensive result record
+                    result = self._create_result_record(row, predicted_label, is_correct, response.Reason)
+                    
+                    tqdm.write(f"Record {i+1} - Actual: {'Scam' if actual_label == 1 else 'Legitimate'}, "
+                              f"Predicted: {'Scam' if predicted_label == 1 else 'Legitimate'}, Correct: {is_correct}")
+                    
+                    return result
                     
                 except Exception as e:
-                    tqdm.write(f"  Error annotating record {i+1}: {e}")
-                    annotation = self._create_error_annotation_record(i+1, row, str(e))
-                    return annotation
+                    tqdm.write(f"  Error processing record {i+1}: {e}")
+                    result = self._create_error_result_record(row, str(e))
+                    return result
                 finally:
                     # Update progress bar after each record
                     pbar.update(1)
@@ -335,259 +271,218 @@ class LLMAnnotationPipeline:
         # Create tasks for all records
         tasks = []
         for i, (_, row) in enumerate(sample_df.iterrows()):
-            task = annotate_single_record(i, row)
+            task = evaluate_single_record(i, row)
             tasks.append(task)
         
         # Execute all tasks concurrently
         start_time = time.time()
-        annotations = await asyncio.gather(*tasks, return_exceptions=False)
+        results = await asyncio.gather(*tasks, return_exceptions=False)
         end_time = time.time()
         
         # Close progress bar
         pbar.close()
         
         # Handle any exceptions that might have been returned
-        valid_annotations = []
-        for i, annotation in enumerate(annotations):
-            if isinstance(annotation, Exception):
-                tqdm.write(f"Exception in record {i+1}: {annotation}")
+        valid_results = []
+        for i, result in enumerate(results):
+            if isinstance(result, Exception):
+                tqdm.write(f"Exception in record {i+1}: {result}")
                 # Create error record
                 row = sample_df.iloc[i]
-                error_annotation = self._create_error_annotation_record(i+1, row, str(annotation))
-                valid_annotations.append(error_annotation)
+                error_result = self._create_error_result_record(row, str(result))
+                valid_results.append(error_result)
             else:
-                valid_annotations.append(annotation)
+                valid_results.append(result)
         
-        self.annotations = valid_annotations
+        self.results = valid_results
         total_time = end_time - start_time
-        print(f"\nAsync annotation completed in {total_time:.2f} seconds. Processed {len(valid_annotations)} records.")
-        print(f"Average time per record: {total_time/len(valid_annotations):.2f} seconds")
-        return valid_annotations
+        print(f"\nAsync evaluation completed in {total_time:.2f} seconds. Processed {len(valid_results)} records.")
+        print(f"Average time per record: {total_time/len(valid_results):.2f} seconds")
+        return valid_results
     
-    def _get_annotation_prompt_template(self):
-        """Get prompt template for annotations"""
-        from langchain_core.prompts import ChatPromptTemplate
-        return ChatPromptTemplate([
-            ("system", "{system_prompt}"),
-            ("user", "{user_prompt}")
-        ])
-    
-    def _create_annotation_record(self, 
-                                 record_id: int, 
-                                 row: pd.Series, 
-                                 response: AnnotationResponseSchema) -> Dict[str, Any]:
-        """Create a comprehensive annotation record"""
-        annotation = {
-            'record_id': record_id,
-            'label': row['label'],
-            'class': 'Scam' if row['label'] == 1 else 'Legitimate',
-            'explanation': response.explanation,
-            'key_indicators': response.key_indicators,
-            'confidence': response.confidence,
-            'usability': response.usability,
-            'annotation_timestamp': pd.Timestamp.now().isoformat()
+    def _create_result_record(self, 
+                             row: pd.Series, 
+                             predicted_label: int, 
+                             is_correct: bool, 
+                             llm_reason: str) -> Dict[str, Any]:
+        """Create a comprehensive result record including original data"""
+        # Label should be clean at this point due to pre-filtering
+        actual_label = int(row['label'])
+        
+        result = {
+            'actual_label': actual_label,
+            'actual_class': 'Scam' if actual_label == 1 else 'Legitimate',
+            'predicted_label': predicted_label,
+            'predicted_class': 'Scam' if predicted_label == 1 else 'Legitimate',
+            'is_correct': is_correct,
+            'llm_reason': llm_reason
         }
         
         # Add id column if it exists
         if 'id' in row:
-            annotation['id'] = row['id']
+            result['id'] = row['id']
         
         # Add all original features except id (to avoid duplication)
         for feature in self.data_loader.features:
             if feature != 'id':
-                annotation[f'original_{feature}'] = row[feature]
+                result[f'original_{feature}'] = row[feature]
         
-        return annotation
+        return result
     
-    def _create_error_annotation_record(self, record_id: int, row: pd.Series, error_message: str) -> Dict[str, Any]:
-        """Create an error annotation record"""
-        annotation = {
-            'record_id': record_id,
-            'label': row['label'],
-            'class': 'Scam' if row['label'] == 1 else 'Legitimate',
-            'explanation': f'Error during annotation: {error_message}',
-            'key_indicators': [],
-            'confidence': 'error',
-            'usability': 0,  # Default to not usable for error cases
-            'annotation_timestamp': pd.Timestamp.now().isoformat()
+    def _create_error_result_record(self, row: pd.Series, error_message: str) -> Dict[str, Any]:
+        """Create an error result record"""
+        # Label should be clean at this point due to pre-filtering
+        actual_label = int(row['label'])
+        
+        result = {
+            'actual_label': actual_label,
+            'actual_class': 'Scam' if actual_label == 1 else 'Legitimate',
+            'predicted_label': None,
+            'predicted_class': 'Error',
+            'is_correct': False,
+            'llm_reason': f'Error: {error_message}'
         }
         
         # Add id column if it exists
         if 'id' in row:
-            annotation['id'] = row['id']
+            result['id'] = row['id']
         
         # Add all original features except id (to avoid duplication)
         for feature in self.data_loader.features:
             if feature != 'id':
-                annotation[f'original_{feature}'] = row[feature]
+                result[f'original_{feature}'] = row[feature]
         
-        return annotation
+        return result
     
-    def save_annotations(self) -> Dict[str, str]:
-        """Save annotation results in result/annotated/{dataset_name}/{timestamp}/ structure"""
-        if not self.annotations:
-            raise ValueError("No annotations available. Run annotation first.")
+    def calculate_metrics(self) -> Dict[str, Any]:
+        """Calculate performance metrics"""
+        if not self.results:
+            raise ValueError("No results available. Run evaluation first.")
         
-        # Create directory structure: result/annotated/{dataset_name}/{timestamp}/
-        timestamp = pd.Timestamp.now().strftime("%Y%m%d_%H%M%S")
-        results_dir = Path(self.output_dir) / self.data_loader.dataset_name / timestamp
-        results_dir.mkdir(parents=True, exist_ok=True)
+        calculator = MetricsCalculator(self.results)
+        metrics = calculator.calculate_metrics()
         
-        # Save annotated dataset
-        annotations_df = pd.DataFrame(self.annotations)
-        annotations_path = results_dir / f"{self.data_loader.dataset_name}_annotated.csv"
-        annotations_df.to_csv(annotations_path, index=False)
+        # Print summary
+        dataset_info = self.data_loader.get_dataset_info()
+        dataset_info['features'] = self.content_columns
+        #calculator.print_metrics_summary(dataset_info)
         
-        # Create and save annotation summary
-        summary_path = self._save_annotation_summary(results_dir)
-        
-        print(f"\nAnnotation results saved to: {results_dir}")
-        print(f"- Annotated dataset: {annotations_path}")
-        print(f"- Summary: {summary_path}")
-        
-        return {
-            'results_directory': str(results_dir),
-            'annotated_dataset_path': str(annotations_path),
-            'summary_path': str(summary_path)
-        }
+        return metrics
     
-    def _save_annotation_summary(self, results_dir: Path) -> Path:
-        """Create and save annotation summary"""
-        if not self.annotations:
-            return results_dir / "annotation_summary.json"
+    def save_results(self) -> Dict[str, str]:
+        """Save results to the specified directory structure"""
+        if not self.results:
+            raise ValueError("No results available. Run evaluation first.")
         
-        # Calculate annotation statistics
-        total_annotations = len(self.annotations)
-        successful_annotations = len([a for a in self.annotations if a['confidence'] != 'error'])
-        error_count = total_annotations - successful_annotations
-        
-        # Confidence distribution
-        confidence_dist = {}
-        for annotation in self.annotations:
-            conf = annotation['confidence']
-            confidence_dist[conf] = confidence_dist.get(conf, 0) + 1
-        
-        # Class distribution
-        class_dist = {}
-        for annotation in self.annotations:
-            cls = annotation['class']
-            class_dist[cls] = class_dist.get(cls, 0) + 1
-        
-        # Usability distribution
-        usability_dist = {}
-        for annotation in self.annotations:
-            usability = annotation.get('usability', 0)
-            usability_key = 'Usable' if usability == 1 else 'Not usable'
-            usability_dist[usability_key] = usability_dist.get(usability_key, 0) + 1
-        
-        # Average indicators per class
-        scam_indicators = [len(a['key_indicators']) for a in self.annotations if a['class'] == 'Scam' and a['confidence'] != 'error']
-        legit_indicators = [len(a['key_indicators']) for a in self.annotations if a['class'] == 'Legitimate' and a['confidence'] != 'error']
-        
-        # Get dataset info
+        # Calculate metrics
+        metrics = self.calculate_metrics()
         dataset_info = self.data_loader.get_dataset_info()
         
-        summary = {
-            'annotation_info': {
-                'timestamp': pd.Timestamp.now().isoformat(),
-                'dataset_name': self.data_loader.dataset_name,
-                'dataset_path': self.dataset_path,
-                'model_provider': self.provider,
-                'model_name': self.model,
-                'sample_size_requested': self.sample_size,
-                'balanced_sample': self.balanced_sample,
-                'random_state': self.random_state,
-                'content_columns': self.content_columns
-            },
-            'dataset_info': {
-                'total_records_in_dataset': dataset_info['total_records'],
-                'scam_count_in_dataset': dataset_info['scam_count'],
-                'legitimate_count_in_dataset': dataset_info['legitimate_count'],
-                'all_features': dataset_info['features']
-            },
-            'annotation_results': {
-                'total_annotations': total_annotations,
-                'successful_annotations': successful_annotations,
-                'error_count': error_count,
-                'success_rate': successful_annotations / total_annotations if total_annotations > 0 else 0
-            },
-            'confidence_distribution': confidence_dist,
-            'class_distribution': class_dist,
-            'usability_distribution': usability_dist,
-            'indicator_statistics': {
-                'avg_indicators_scam': sum(scam_indicators) / len(scam_indicators) if scam_indicators else 0,
-                'avg_indicators_legitimate': sum(legit_indicators) / len(legit_indicators) if legit_indicators else 0,
-                'max_indicators': max([len(a['key_indicators']) for a in self.annotations if a['confidence'] != 'error'], default=0),
-                'min_indicators': min([len(a['key_indicators']) for a in self.annotations if a['confidence'] != 'error'], default=0)
-            }
-        }
+        dataset_info['features'] = self.content_columns
+        dataset_info['features_used'] = self.content_columns
+        calculator = MetricsCalculator(self.results)
+        calculator.print_metrics_summary(dataset_info)
+        # Initialize results saver
+        saver = ResultsSaver(
+            dataset_name=self.data_loader.dataset_name,
+            provider=self.provider,
+            model=self.model
+        )
         
-        # Save summary
-        import json
-        summary_path = results_dir / "annotation_summary.json"
-        with open(summary_path, 'w') as f:
-            json.dump(summary, f, indent=2)
+        # Save results
+        save_paths = saver.save_results(self.results, metrics, dataset_info)
         
-        print(f"\nAnnotation Summary:")
-        print(f"- Total records: {total_annotations}")
-        print(f"- Successful annotations: {successful_annotations}")
-        print(f"- Success rate: {summary['annotation_results']['success_rate']:.2%}")
-        print(f"- Usable records: {usability_dist.get('Usable', 0)}")
-        print(f"- Not usable records: {usability_dist.get('Not usable', 0)}")
-        print(f"- Usability rate: {usability_dist.get('Usable', 0) / total_annotations:.2%}" if total_annotations > 0 else "- Usability rate: 0%")
+        # Create summary report
+        report_path = saver.create_summary_report(metrics, dataset_info)
+        save_paths['summary_report_path'] = report_path
         
-        return summary_path
+        return save_paths
     
-    def run_full_annotation(self) -> Dict[str, Any]:
-        """Run the complete annotation pipeline"""
-        print("Starting LLM Annotation Pipeline...")
+    def run_full_evaluation(self) -> Dict[str, Any]:
+        """
+        Run complete evaluation pipeline: setup -> load -> evaluate -> calculate -> save
+        
+        Returns:
+            Dictionary with evaluation results and file paths
+        """
+        print("="*80)
+        print("SCAM DETECTION EVALUATION PIPELINE")
+        print("="*80)
         
         # Setup LLM
+        print("\n1. Setting up LLM...")
         self.setup_llm()
         
         # Load and prepare data
+        print("\n2. Loading and preparing data...")
         sample_df = self.load_and_prepare_data()
         
-        # Run annotation
-        annotations = self.annotate_sample(sample_df)
+        # Run evaluation
+        print("\n3. Running evaluation...")
+        results = self.evaluate_sample(sample_df)
+        
+        # Calculate metrics
+        print("\n4. Calculating metrics...")
+        metrics = self.calculate_metrics()
         
         # Save results
-        save_paths = self.save_annotations()
+        print("\n5. Saving results...")
+        save_paths = self.save_results()
+        
+        print("\n" + "="*80)
+        print("EVALUATION COMPLETED SUCCESSFULLY")
+        print("="*80)
         
         return {
-            'annotations': annotations,
+            'results': results,
+            'metrics': metrics,
             'save_paths': save_paths,
-            'summary': {
-                'total_records': len(annotations),
-                'successful_annotations': len([a for a in annotations if a['confidence'] != 'error']),
-                'dataset_name': self.data_loader.dataset_name
-            }
+            'dataset_info': self.data_loader.get_dataset_info()
         }
 
-    async def run_full_annotation_async(self, concurrent_requests: int = 10) -> Dict[str, Any]:
-        """Run the complete annotation pipeline asynchronously"""
-        print("Starting LLM Annotation Pipeline (Async)...")
+    async def run_full_evaluation_async(self, concurrent_requests: int = 10) -> Dict[str, Any]:
+        """
+        Run complete evaluation pipeline asynchronously: setup -> load -> evaluate -> calculate -> save
+        
+        Args:
+            concurrent_requests: Number of concurrent requests to make (default: 10)
+            
+        Returns:
+            Dictionary with evaluation results and file paths
+        """
+        print("="*80)
+        print("SCAM DETECTION EVALUATION PIPELINE (ASYNC)")
+        print("="*80)
         
         # Setup LLM
+        print("\n1. Setting up LLM...")
         self.setup_llm()
         
         # Load and prepare data
+        print("\n2. Loading and preparing data...")
         sample_df = self.load_and_prepare_data()
         
-        # Run annotation asynchronously
-        annotations = await self.annotate_sample_async(sample_df, concurrent_requests)
+        # Run evaluation asynchronously
+        print("\n3. Running evaluation asynchronously...")
+        results = await self.evaluate_sample_async(sample_df, concurrent_requests)
+        
+        # Calculate metrics
+        print("\n4. Calculating metrics...")
+        metrics = self.calculate_metrics()
         
         # Save results
-        save_paths = self.save_annotations()
+        print("\n5. Saving results...")
+        save_paths = self.save_results()
+        
+        print("\n" + "="*80)
+        print("ASYNC EVALUATION COMPLETED SUCCESSFULLY")
+        print("="*80)
         
         return {
-            'annotations': annotations,
+            'results': results,
+            'metrics': metrics,
             'save_paths': save_paths,
-            'summary': {
-                'total_records': len(annotations),
-                'successful_annotations': len([a for a in annotations if a['confidence'] != 'error']),
-                'dataset_name': self.data_loader.dataset_name
-            }
+            'dataset_info': self.data_loader.get_dataset_info()
         }
     
     # ==================== CHECKPOINT FUNCTIONALITY ====================
@@ -598,7 +493,7 @@ class LLMAnnotationPipeline:
         # Sanitize model name for filesystem compatibility
         safe_model_name = self.model.replace("/", "-").replace("\\", "-").replace(":", "-")
         timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        return f"{dataset_name}_annotation_{self.provider}_{safe_model_name}_{timestamp}.json"
+        return f"{dataset_name}_evaluation_{self.provider}_{safe_model_name}_{timestamp}.json"
     
     def _find_existing_checkpoint(self, checkpoint_dir: str = "checkpoints") -> Optional[Path]:
         """Find the most recent checkpoint file for this configuration"""
@@ -609,7 +504,7 @@ class LLMAnnotationPipeline:
         dataset_name = Path(self.dataset_path).stem
         # Sanitize model name for filesystem compatibility
         safe_model_name = self.model.replace("/", "-").replace("\\", "-").replace(":", "-")
-        pattern = f"{dataset_name}_annotation_{self.provider}_{safe_model_name}_*.json"
+        pattern = f"{dataset_name}_evaluation_{self.provider}_{safe_model_name}_*.json"
         
         checkpoint_files = list(checkpoint_path.glob(pattern))
         if checkpoint_files:
@@ -649,11 +544,11 @@ class LLMAnnotationPipeline:
             
             # Load checkpoint state
             self.current_index = checkpoint_data.get('current_index', 0)
-            self.annotations = checkpoint_data.get('annotations', [])
+            self.results = checkpoint_data.get('results', [])
             self.checkpoint_file = checkpoint_file
             
             print(f"Loaded checkpoint from {checkpoint_file}")
-            print(f"Resuming from index {self.current_index} with {len(self.annotations)} existing annotations")
+            print(f"Resuming from index {self.current_index} with {len(self.results)} existing results")
             return True
             
         except Exception as e:
@@ -679,7 +574,7 @@ class LLMAnnotationPipeline:
             'use_structure_model': self.use_structure_model,
             'current_index': self.current_index,
             'total_records': self.total_records,
-            'annotations': self.annotations,
+            'results': self.results,
             'timestamp': datetime.now().isoformat(),
             'elapsed_time': time.time() - self.start_time if self.start_time else 0
         }
@@ -690,16 +585,17 @@ class LLMAnnotationPipeline:
         except Exception as e:
             print(f"Error saving checkpoint: {e}")
     
-    def _write_checkpoint_message(self, message: str):
-        """Write checkpoint message, clearing the previous one"""
-        # Clear previous checkpoint message if it exists
-        if self.last_checkpoint_message:
-            # Move cursor up and clear line
-            tqdm.write('\033[F\033[K', end='')
-        
-        # Write new checkpoint message
-        tqdm.write(message)
+    def _write_checkpoint_message(self, message: str, progress_bar=None):
+        """Write checkpoint message using tqdm-friendly approach"""
+        # Use tqdm.write() which automatically positions the message correctly
+        tqdm.write(f"📁 {message}")
         self.last_checkpoint_message = message
+        
+        # If we have a progress bar reference, update its postfix
+        if progress_bar:
+            progress_bar.set_postfix_str("✓ Checkpoint saved")
+            # Clear the postfix after a brief moment to keep it clean
+            progress_bar.refresh()
     
     def process_full_dataset_with_checkpoints(self, 
                                             checkpoint_interval: int = 1000,
@@ -716,10 +612,10 @@ class LLMAnnotationPipeline:
             override_compatibility: Whether to override checkpoint compatibility checks
             
         Returns:
-            Dictionary with annotation results and file paths
+            Dictionary with evaluation results and file paths
         """
         print("="*80)
-        print("ANNOTATION PIPELINE WITH CHECKPOINTING")
+        print("EVALUATION PIPELINE WITH CHECKPOINTING")
         print("="*80)
         
         # Setup LLM
@@ -739,7 +635,7 @@ class LLMAnnotationPipeline:
             self.content_columns = self.data_loader.features
         
         # Initialize prompt generator
-        self.prompt_generator = AnnotationPromptGenerator(self.data_loader.features, self.content_columns)
+        self.prompt_generator = PromptGenerator(self.data_loader.features, self.content_columns)
         
         print(f"Dataset: {self.data_loader.dataset_name}")
         print(f"Total records: {self.total_records:,}")
@@ -750,14 +646,17 @@ class LLMAnnotationPipeline:
         self.load_checkpoint(checkpoint_dir, resume_from_checkpoint, override_compatibility)
         
         self.start_time = time.time()
-        total_processed = len(self.annotations)
+        total_processed = len(self.results)
         
         print(f"\nProcessing records {self.current_index + 1} to {self.total_records}...")
+        
+        # Get system prompt
+        system_prompt = self.prompt_generator.get_system_prompt()
         
         # Create progress bar
         progress_bar = tqdm(
             range(self.current_index, self.total_records),
-            desc="Annotating",
+            desc="Evaluating",
             unit="records",
             initial=self.current_index,
             total=self.total_records
@@ -768,12 +667,11 @@ class LLMAnnotationPipeline:
             row = dataset_df.iloc[i]
             
             # Update progress bar description
-            progress_bar.set_description(f"Annotating record {i + 1}/{self.total_records}")
+            progress_bar.set_description(f"Evaluating record {i + 1}/{self.total_records}")
             
             try:
-                # Create annotation prompt
-                system_prompt = self.prompt_generator.get_system_prompt()
-                user_prompt = self.prompt_generator.create_annotation_prompt(row.to_dict(), str(row['label']))
+                # Create user prompt
+                user_prompt = self.prompt_generator.create_user_prompt(row.to_dict())
                 
                 # Make API call
                 if self.use_structure_model:
@@ -782,29 +680,31 @@ class LLMAnnotationPipeline:
                     if self.enable_thinking or ('<think>' in response and '</think>' in response):
                         import re
                         response = re.sub(r'<think>.*?</think>', '', response, flags=re.DOTALL).strip()
-                    response = parse_structured_output(self.structure_model, response, AnnotationResponseSchema)
+                    response = parse_structured_output(self.structure_model, response)
                 else:
-                    # Make API call with custom structured output for annotations
-                    prompt_template = self._get_annotation_prompt_template()
-                    messages = prompt_template.invoke({"system_prompt": system_prompt, "user_prompt": user_prompt})
-                    client = self.llm.with_structured_output(AnnotationResponseSchema)
-                    response = client.invoke(messages)
+                    response = make_api_call(self.llm, system_prompt, user_prompt, structure_model=False)
                 
-                # Create annotation record
-                annotation = self._create_annotation_record(i, row, response)
+                # Extract prediction
+                predicted_scam = response.Phishing
+                predicted_label = 1 if predicted_scam else 0
+                actual_label = row['label']
+                is_correct = predicted_label == actual_label
+                
+                # Create evaluation record
+                result = self._create_result_record(row, predicted_label, is_correct, response.Reason)
                 
             except Exception as e:
                 tqdm.write(f"Error processing record {i + 1}: {e}")
-                annotation = self._create_error_annotation_record(i, row, str(e))
+                result = self._create_error_result_record(row, str(e))
             
-            self.annotations.append(annotation)
+            self.results.append(result)
             self.current_index = i + 1
             total_processed += 1
             
             # Save checkpoint at intervals
             if (i + 1) % checkpoint_interval == 0:
                 self.save_checkpoint(checkpoint_dir)
-                self._write_checkpoint_message(f"Checkpoint saved at record {i + 1}")
+                self._write_checkpoint_message(f"Checkpoint saved at record {i + 1}", progress_bar)
         
         # Close progress bar
         progress_bar.close()
@@ -812,24 +712,28 @@ class LLMAnnotationPipeline:
         # Final checkpoint
         self.save_checkpoint(checkpoint_dir)
         
-        # Save final results
-        save_paths = self.save_annotations()
+        # Calculate metrics and save final results
+        metrics = self.calculate_metrics()
+        save_paths = self.save_results()
         
         total_time = time.time() - self.start_time
         print(f"\n{'='*80}")
-        print("ANNOTATION COMPLETED")
+        print("EVALUATION COMPLETED")
         print(f"{'='*80}")
         print(f"Total records processed: {total_processed:,}")
         print(f"Total time: {total_time:.2f} seconds")
         print(f"Average time per record: {total_time/total_processed:.3f} seconds")
+        print(f"Accuracy: {metrics.get('accuracy', 0):.2%}")
         
         return {
-            'annotations': self.annotations,
+            'results': self.results,
+            'metrics': metrics,
             'save_paths': save_paths,
+            'dataset_info': self.data_loader.get_dataset_info(),
             'summary': {
-                'total_records': len(self.annotations),
-                'successful_annotations': len([a for a in self.annotations if a['confidence'] != 'error']),
-                'dataset_name': self.data_loader.dataset_name,
+                'total_records': len(self.results),
+                'successful_evaluations': len([r for r in self.results if r.get('predicted_label') is not None]),
+                'accuracy': metrics.get('accuracy', 0),
                 'checkpoint_file': str(self.checkpoint_file) if self.checkpoint_file else None
             }
         }
@@ -851,10 +755,10 @@ class LLMAnnotationPipeline:
             override_compatibility: Whether to override checkpoint compatibility checks
             
         Returns:
-            Dictionary with annotation results and file paths
+            Dictionary with evaluation results and file paths
         """
         print("="*80)
-        print("ASYNC ANNOTATION PIPELINE WITH CHUNKED BATCHING")
+        print("ASYNC EVALUATION PIPELINE WITH CHUNKED BATCHING")
         print(f"Concurrent requests: {concurrent_requests}")
         print("="*80)
         
@@ -875,7 +779,7 @@ class LLMAnnotationPipeline:
             self.content_columns = self.data_loader.features
         
         # Initialize prompt generator
-        self.prompt_generator = AnnotationPromptGenerator(self.data_loader.features, self.content_columns)
+        self.prompt_generator = PromptGenerator(self.data_loader.features, self.content_columns)
         
         print(f"Dataset: {self.data_loader.dataset_name}")
         print(f"Total records: {self.total_records:,}")
@@ -886,9 +790,6 @@ class LLMAnnotationPipeline:
         self.load_checkpoint(checkpoint_dir, resume_from_checkpoint, override_compatibility)
         
         self.start_time = time.time()
-        
-        # Create semaphore to limit concurrent requests
-        semaphore = asyncio.Semaphore(concurrent_requests)
         
         print(f"\nProcessing records {self.current_index + 1} to {self.total_records}...")
         
@@ -919,16 +820,16 @@ class LLMAnnotationPipeline:
             )
             
             # Create tasks for this chunk
-            async def annotate_with_semaphore(index: int, row: pd.Series):
+            async def evaluate_with_semaphore(index: int, row: pd.Series):
                 async with semaphore:
-                    result = await self._annotate_single_record_async(index, row)
+                    result = await self._evaluate_single_record_async(index, row)
                     chunk_progress.update(1)
                     return result
             
             tasks = []
             for i in range(chunk_start, chunk_end):
                 row = dataset_df.iloc[i]
-                tasks.append(annotate_with_semaphore(i, row))
+                tasks.append(evaluate_with_semaphore(i, row))
             
             # Process chunk
             chunk_results = await asyncio.gather(*tasks, return_exceptions=True)
@@ -938,9 +839,9 @@ class LLMAnnotationPipeline:
                 actual_index = chunk_start + i
                 if isinstance(result, Exception):
                     row = dataset_df.iloc[actual_index]
-                    result = self._create_error_annotation_record(actual_index, row, str(result))
+                    result = self._create_error_result_record(row, str(result))
                 
-                self.annotations.append(result)
+                self.results.append(result)
                 self.current_index = actual_index + 1
                 overall_progress.update(1)
             
@@ -950,7 +851,7 @@ class LLMAnnotationPipeline:
             # Save checkpoint after each chunk
             if chunk_end % checkpoint_interval <= chunk_size:
                 self.save_checkpoint(checkpoint_dir)
-                self._write_checkpoint_message(f"Checkpoint saved at record {chunk_end}")
+                self._write_checkpoint_message(f"Checkpoint saved at record {chunk_end}", overall_progress)
         
         # Close overall progress bar
         overall_progress.close()
@@ -958,35 +859,39 @@ class LLMAnnotationPipeline:
         # Final checkpoint
         self.save_checkpoint(checkpoint_dir)
         
-        # Save final results
-        save_paths = self.save_annotations()
+        # Calculate metrics and save final results
+        metrics = self.calculate_metrics()
+        save_paths = self.save_results()
         
         total_time = time.time() - self.start_time
-        total_processed = len(self.annotations)
+        total_processed = len(self.results)
         
         print(f"\n{'='*80}")
-        print("CHUNKED ASYNC ANNOTATION COMPLETED")
+        print("CHUNKED ASYNC EVALUATION COMPLETED")
         print(f"{'='*80}")
         print(f"Total records processed: {total_processed:,}")
         print(f"Total time: {total_time:.2f} seconds")
         print(f"Average time per record: {total_time/total_processed:.3f} seconds")
+        print(f"Accuracy: {metrics.get('accuracy', 0):.2%}")
         
         return {
-            'annotations': self.annotations,
+            'results': self.results,
+            'metrics': metrics,
             'save_paths': save_paths,
+            'dataset_info': self.data_loader.get_dataset_info(),
             'summary': {
-                'total_records': len(self.annotations),
-                'successful_annotations': len([a for a in self.annotations if a['confidence'] != 'error']),
-                'dataset_name': self.data_loader.dataset_name,
+                'total_records': len(self.results),
+                'successful_evaluations': len([r for r in self.results if r.get('predicted_label') is not None]),
+                'accuracy': metrics.get('accuracy', 0),
                 'checkpoint_file': str(self.checkpoint_file) if self.checkpoint_file else None
             }
         }
     
-    async def _annotate_single_record_async(self, index: int, row: pd.Series) -> Dict[str, Any]:
-        """Annotate a single record asynchronously"""
+    async def _evaluate_single_record_async(self, index: int, row: pd.Series) -> Dict[str, Any]:
+        """Evaluate a single record asynchronously"""
         try:
             system_prompt = self.prompt_generator.get_system_prompt()
-            user_prompt = self.prompt_generator.create_annotation_prompt(row.to_dict(), str(row['label']))
+            user_prompt = self.prompt_generator.create_user_prompt(row.to_dict())
             
             if self.use_structure_model:
                 response = await make_api_call_async(self.llm, system_prompt, user_prompt,
@@ -994,18 +899,19 @@ class LLMAnnotationPipeline:
                 if self.enable_thinking or ('<think>' in response and '</think>' in response):
                     import re
                     response = re.sub(r'<think>.*?</think>', '', response, flags=re.DOTALL).strip()
-                response = await parse_structured_output_async(self.structure_model, response, AnnotationResponseSchema)
+                response = await parse_structured_output_async(self.structure_model, response)
             else:
-                # Make async API call with custom structured output for annotations
-                prompt_template = self._get_annotation_prompt_template()
-                messages = prompt_template.invoke({"system_prompt": system_prompt, "user_prompt": user_prompt})
-                client = self.llm.with_structured_output(AnnotationResponseSchema)
-                response = await client.ainvoke(messages)
+                response = await make_api_call_async(self.llm, system_prompt, user_prompt, structure_model=False)
             
-            return self._create_annotation_record(index, row, response)
+            predicted_scam = response.Phishing
+            predicted_label = 1 if predicted_scam else 0
+            actual_label = row['label']
+            is_correct = predicted_label == actual_label
+            
+            return self._create_result_record(row, predicted_label, is_correct, response.Reason)
             
         except Exception as e:
-            return self._create_error_annotation_record(index, row, str(e))
+            return self._create_error_result_record(row, str(e))
 
     async def process_full_dataset_with_checkpoints_async(self, 
                                                          checkpoint_interval: int = 1000,
@@ -1027,10 +933,10 @@ class LLMAnnotationPipeline:
             override_compatibility: Whether to override checkpoint compatibility checks
             
         Returns:
-            Dictionary with annotation results and file paths
+            Dictionary with evaluation results and file paths
         """
         print("="*80)
-        print("ASYNC ANNOTATION PIPELINE WITH OVERLAPPING BATCHES")
+        print("ASYNC EVALUATION PIPELINE WITH OVERLAPPING BATCHES")
         print(f"Concurrent requests: {concurrent_requests}")
         print("="*80)
         
@@ -1051,7 +957,7 @@ class LLMAnnotationPipeline:
             self.content_columns = self.data_loader.features
         
         # Initialize prompt generator
-        self.prompt_generator = AnnotationPromptGenerator(self.data_loader.features, self.content_columns)
+        self.prompt_generator = PromptGenerator(self.data_loader.features, self.content_columns)
         
         print(f"Dataset: {self.data_loader.dataset_name}")
         print(f"Total records: {self.total_records:,}")
@@ -1099,7 +1005,7 @@ class LLMAnnotationPipeline:
         async def process_single_record(index: int, row: pd.Series):
             """Process a single record with semaphore control"""
             async with semaphore:
-                return await self._annotate_single_record_async(index, row)
+                return await self._evaluate_single_record_async(index, row)
         
         # Main processing loop with overlapping batches
         while next_index < self.total_records or active_tasks:
@@ -1125,7 +1031,7 @@ class LLMAnnotationPipeline:
                         result = await task
                     except Exception as e:
                         row = dataset_df.iloc[index]
-                        result = self._create_error_annotation_record(index, row, str(e))
+                        result = self._create_error_result_record(row, str(e))
                     
                     completed_results[index] = result
                     overall_progress.update(1)
@@ -1133,7 +1039,7 @@ class LLMAnnotationPipeline:
                 # Process completed results in order and add to final results
                 while self.current_index in completed_results:
                     result = completed_results.pop(self.current_index)
-                    self.annotations.append(result)
+                    self.results.append(result)
                     self.current_index += 1
                     
                     # Enhanced time estimation updates every 100 records
@@ -1210,6 +1116,9 @@ class LLMAnnotationPipeline:
                     if self.current_index % checkpoint_interval == 0:
                         self.save_checkpoint(checkpoint_dir)
                         
+                        # Calculate session progress for checkpoint messaging
+                        records_processed_this_session = self.current_index - session_start_index
+                        
                         # Update checkpoint with enhanced timing data
                         try:
                             checkpoint_file = Path(checkpoint_dir) / f"{self.checkpoint_file}.json"
@@ -1232,7 +1141,8 @@ class LLMAnnotationPipeline:
                         elapsed_session = time.time() - self.start_time
                         self._write_checkpoint_message(
                             f"✓ Checkpoint saved at record {self.current_index:,} "
-                            f"(session: {elapsed_session/60:.1f}m, rate: {records_processed_this_session/elapsed_session:.2f}r/s)"
+                            f"(session: {elapsed_session/60:.1f}m, rate: {records_processed_this_session/elapsed_session:.2f}r/s)",
+                            overall_progress
                         )
         
         # Close progress bar
@@ -1247,7 +1157,7 @@ class LLMAnnotationPipeline:
         print(f"   Session time: {total_time:.2f} seconds ({total_time/3600:.2f} hours)")
         print(f"   Session rate: {final_rate:.2f} records/second")
         print(f"   Records processed this session: {records_processed_this_session:,}")
-        print(f"   Total records completed: {len(self.annotations):,}")
+        print(f"   Total records completed: {len(self.results):,}")
         
         if rates_history:
             print(f"   Recent rate: {rates_history[-1]:.2f} records/second")
@@ -1256,192 +1166,33 @@ class LLMAnnotationPipeline:
         # Final checkpoint
         self.save_checkpoint(checkpoint_dir)
         
-        # Save final results
-        save_paths = self.save_annotations()
+        # Calculate metrics and save final results
+        metrics = self.calculate_metrics()
+        save_paths = self.save_results()
         
         print(f"\n{'='*80}")
-        print("ASYNC ANNOTATION COMPLETED")
+        print("ASYNC EVALUATION COMPLETED")
         print(f"{'='*80}")
-        print(f"Total records completed: {len(self.annotations):,}")
+        print(f"Total records completed: {len(self.results):,}")
         print(f"Session processing rate: {final_rate:.2f} records/second")
+        print(f"Accuracy: {metrics.get('accuracy', 0):.2%}")
         
         return {
-            'annotations': self.annotations,
+            'results': self.results,
+            'metrics': metrics,
             'save_paths': save_paths,
+            'dataset_info': self.data_loader.get_dataset_info(),
             'summary': {
-                'total_records': len(self.annotations),
-                'successful_annotations': len([a for a in self.annotations if a['confidence'] != 'error']),
-                'usable_annotations': len([a for a in self.annotations if a.get('usability', True)]),
-                'success_rate': len([a for a in self.annotations if a['confidence'] != 'error']) / len(self.annotations),
-                'usability_rate': len([a for a in self.annotations if a.get('usability', True)]) / len(self.annotations),
-                'dataset_name': self.data_loader.dataset_name,
+                'total_records': len(self.results),
+                'successful_evaluations': len([r for r in self.results if r.get('predicted_label') is not None]),
+                'correct_predictions': len([r for r in self.results if r.get('is_correct', False)]),
+                'accuracy': metrics.get('accuracy', 0),
+                'success_rate': len([r for r in self.results if r.get('predicted_label') is not None]) / len(self.results),
                 'final_processing_rate': final_rate,
                 'total_processing_time': total_time,
                 'checkpoint_file': str(self.checkpoint_file) if self.checkpoint_file else None
             }
         }
 
-def parse_arguments():
-    """Parse command line arguments"""
-    parser = argparse.ArgumentParser(
-        description="LLM Annotation Pipeline for Scam Detection",
-        formatter_class=argparse.RawTextHelpFormatter
-    )
-    
-    # Required arguments
-    parser.add_argument(
-        '--dataset', 
-        type=str, 
-        required=True,
-        help='Path to the dataset CSV file (must contain a "label" column)'
-    )
-    
-    parser.add_argument(
-        '--provider', 
-        type=str, 
-        required=True,
-        choices=['openai', 'anthropic', 'gemini', 'local'],
-        help='LLM provider to use'
-    )
-    
-    parser.add_argument(
-        '--model', 
-        type=str, 
-        required=True,
-        help='Model name to use (e.g., gpt-4, claude-3-sonnet, gemini-pro, etc.)'
-    )
-    
-    # Optional arguments
-    parser.add_argument(
-        '--sample-size', 
-        type=int, 
-        default=100,
-        help='Number of samples to annotate (default: 100)'
-    )
-    
-    parser.add_argument(
-        '--balanced-sample',
-        action='store_true',
-        help='Sample equal numbers of scam and legitimate messages'
-    )
-    
-    parser.add_argument(
-        '--random-state', 
-        type=int, 
-        default=42,
-        help='Random seed for reproducibility (default: 42)'
-    )
-    
-    parser.add_argument(
-        '--content-columns',
-        type=str,
-        nargs='+',
-        help='Specific columns to use as content (e.g., --content-columns subject body)'
-    )
-    
-    parser.add_argument(
-        '--output-dir',
-        type=str,
-        default='annotations',
-        help='Directory for output results (default: annotations)'
-    )
-    
-    parser.add_argument(
-        '--enable-thinking',
-        action='store_true',
-        help='Enable thinking tokens in prompts'
-    )
-    
-    parser.add_argument(
-        '--use-structure-model',
-        action='store_true',
-        help='Use a separate structure model for parsing responses'
-    )
-    
-    return parser.parse_args()
-
-def validate_dataset(dataset_path: str):
-    """Validate that the dataset exists and has required structure"""
-    if not Path(dataset_path).exists():
-        raise FileNotFoundError(f"Dataset file not found: {dataset_path}")
-    
-    import pandas as pd
-    try:
-        df_sample = pd.read_csv(dataset_path, nrows=1)
-        if 'label' not in df_sample.columns:
-            raise ValueError("Dataset must contain a 'label' column")
-        print(f"✓ Dataset validation passed")
-        print(f"  - File: {dataset_path}")
-        print(f"  - Columns: {list(df_sample.columns)}")
-    except Exception as e:
-        raise ValueError(f"Invalid dataset format: {e}")
-
-def main():
-    """Main annotation pipeline execution"""
-    try:
-        # Parse arguments
-        args = parse_arguments()
-        
-        print("="*80)
-        print("LLM ANNOTATION PIPELINE FOR SCAM DETECTION")
-        print("="*80)
-        print(f"Dataset: {args.dataset}")
-        print(f"Provider: {args.provider}")
-        print(f"Model: {args.model}")
-        print(f"Sample size: {args.sample_size}")
-        print(f"Balanced sampling: {args.balanced_sample}")
-        print(f"Random state: {args.random_state}")
-        if args.content_columns:
-            print(f"Content columns: {args.content_columns}")
-        else:
-            print(f"Content columns: All non-label columns")
-        print(f"Output directory: {args.output_dir}")
-        print(f"Enable thinking: {args.enable_thinking}")
-        print(f"Use structure model: {args.use_structure_model}")
-        
-        # Validate dataset
-        print(f"\nValidating dataset...")
-        validate_dataset(args.dataset)
-        
-        # Initialize annotation pipeline
-        pipeline = LLMAnnotationPipeline(
-            dataset_path=args.dataset,
-            provider=args.provider,
-            model=args.model,
-            sample_size=args.sample_size,
-            balanced_sample=args.balanced_sample,
-            random_state=args.random_state,
-            content_columns=args.content_columns,
-            output_dir=args.output_dir,
-            enable_thinking=args.enable_thinking,
-            use_structure_model=args.use_structure_model
-        )
-        
-        # Run full annotation
-        results = pipeline.run_full_annotation()
-        
-        # Print final summary
-        print(f"\n{'='*80}")
-        print("ANNOTATION COMPLETED")
-        print(f"{'='*80}")
-        
-        save_paths = results['save_paths']
-        summary = results['summary']
-        
-        print(f"Results directory: {save_paths['results_directory']}")
-        print(f"Detailed annotations: {save_paths['detailed_results_path']}")
-        print(f"Total records annotated: {summary['total_records']}")
-        print(f"Successful annotations: {summary['successful_annotations']}")
-        
-        print(f"\n✓ Annotation pipeline completed successfully!")
-        
-    except KeyboardInterrupt:
-        print("\n\nAnnotation interrupted by user.")
-        sys.exit(1)
-        
-    except Exception as e:
-        print(f"\n❌ Error: {e}")
-        sys.exit(1)
-
-if __name__ == "__main__":
-    main() 
+# Backward compatibility alias
+PhishingEvaluator = ScamDetectionEvaluator
