+++ src/llm_core/__init__.py
@@ -1,0 +1,12 @@
+from .api_provider import LLM
+from .api_call import (
+    make_api_call, 
+    make_api_call_async,
+    parse_structured_output,
+    parse_structured_output_async,
+    response_schema
+)
+
+__all__ = ['LLM', 'make_api_call', 'make_api_call_async', 
+           'parse_structured_output', 'parse_structured_output_async', 
+           'response_schema']
