+++ src/synthesize/__init__.py
@@ -1,0 +1,4 @@
+from .transcript_generator import TranscriptGenerator
+from .transcript_prompts import get_model_config, get_prompt_for_category
+
+__all__ = ['TranscriptGenerator', 'get_model_config', 'get_prompt_for_category']
