+++ src/synthesize/transcript_prompts.py
